Move connect components to 'i'm new' page
