+++ frontend/src/components/page-new/map-details.js
