+++ frontend/src/pages/new/index.js
@@ -3,10 +3,11 @@ import * as React from "react";
 import Layout from "../../components/layout";
 import Seo, { PageDescriptions } from "../../components/seo";
 import WelcomeSection from "../../components/page-new/welcome";
-import SundayCelebrationSection from "../../components/page-new/sunday-celebration";
 import Questions from "../../components/gotQuestions";
 import Banner from "../../components/shared/banner";
 import LifeStages from "../../components/page-new/lifeStages";
+import CommonQuestions from "../../components/page-new/common-questions";
+import MapDetails from "../../components/page-new/map-details";
 
 const NewcomersPage = () => (
   <Layout>
@@ -14,7 +15,8 @@ const NewcomersPage = () => (
     <div className="content-padding-full">
       <WelcomeSection />
     </div>
-    <SundayCelebrationSection />
+    <MapDetails />
+    <CommonQuestions />
     <LifeStages />
     <div className="pl-[21px] pr-[11px] lg:px-4 w-full flex justify-center">
       <Questions />
