+++ frontend/src/pages/new/life-stages.js
